refactor: simplify platform support in starter package
